// packages/lightroom/src/flat/types.gen.ts
export type JobStatusLink = {
    self?: SelfLink;
};
// ...
export type LrJobApiResponse = {
    /**
     * The job's ID requested.
     */
    jobId?: string;
    /**
     * Created timestamp of the job.
     */
    created?: string;
    /**
     * Modified timestamp of the job.
     */
    modified?: string;
    /**
     * Output parameters for the job.
     */
    outputs?: Array<JobOutputDetails>;
    _links?: JobStatusLink;
};
// ...
export type Links = {
    self?: StorageDetails;
};
// ...
export type JobOutputDetails = {
    /**
     * The original input file path.
     */
    input?: string;
    status?: JobStatus;
    _links?: Links;
    /**
     * Description of the exact error that is reported by the requested job. Will appear in the response only if there is an error.
     */
    details?: string;
};

/**
 * SelfLink
 *
 * Self link for the job status.
 */
export type SelfLink = {
    /**
     * Job Status URL.
     */
    href?: string;
};

/**
 * JobStatus
 *
 * Status of the job.
 */
export type JobStatus = 'pending' | 'running' | 'succeeded' | 'failed';